Add application description and functionality overview
This PHP Laravel App was built to provide a fast food store with an inventory control. The application exposes the products through the UI to allow the service personnel to check what they are delivering and the app estimates the point of reorder according to the historic rotation of the product.